# front/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdoptMe API Documentation",
    "description": "Especificación interactiva de los endpoints de la plataforma de adopción de mascotas AdoptMe. Incluye gestión de usuarios, roles, catálogo de mascotas y solicitudes de adopción.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo local"
    }
  ],
  "paths": {
...
    "/api/user": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "description": "Retorna un listado de todos los usuarios registrados en el sistema.",
        "responses": {
          "200": {
            "description": "Listado de usuarios obtenido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Registrar un usuario nuevo",
        "description": "Crea una nueva cuenta de usuario (usuario común o empresa).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "ejemplo@correo.com" },
                  "password": { "type": "string", "format": "password", "example": "password123" },
                  "nombre": { "type": "string", "example": "Juan Pérez" },
                  "rol": { "type": "string", "enum": ["usuario", "empresa"], "example": "usuario" },
                  "nombreEmpresa": { "type": "string", "description": "Requerido si rol es 'empresa'", "example": "Refugio Patitas" }
                },
                "required": ["email", "password", "rol"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros obligatorios faltantes o tipo de cuenta no válido"
          },
          "409": {
            "description": "El correo electrónico ya está registrado"
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "summary": "Obtener usuario por ID",
        "description": "Retorna la información completa de un usuario específico.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único del usuario",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar usuario",
        "description": "Actualiza la información de perfil (nombre, correo) de un usuario.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único del usuario",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string", "example": "Juan Modificado" },
                  "email": { "type": "string", "format": "email", "example": "nuevo@correo.com" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "El email que intenta usar ya está en uso por otra cuenta"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar usuario",
        "description": "Elimina la cuenta del usuario de la plataforma.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único del usuario",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado correctamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nombre": { "type": "string", "example": "Refugio San Roque" },
          "email": { "type": "string", "format": "email", "example": "sanroque@empresa.com" },
          "tipo_cuenta_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}